refactor: move some code from transfer panel to custom hooks
